sdcsdcs
sdcscdsc

+++ notebooks/test_ui.py
@@ -16,9 +16,9 @@ if uploaded_file is not None:
     #format RSSI_avr
     n_df['RSSI_avr'] = pd.to_numeric(n_df['RSSI_avr'])
     #format Longitude
-    n_df['Longitude'] = n_df['Longitude'].apply(lambda x: x.astype(float))
+    n_df['Longitude'] = n_df['Longitude'].apply(lambda x: float(x))
     #format Latitude
-    n_df['Latitude'] = n_df['Latitude'].apply(lambda x: x.astype(float))
+    n_df['Latitude'] = n_df['Latitude'].apply(lambda x: float(x))
     #format Flag 
     n_df = df.drop(columns=['RSSI_min', 'RSSI_max', 'RSSI_med', 'Flag'])
     #Others
